Ignore overlapping segments
The code was based on code from SponsorBlock relevant to sponsor
segments only. SponsorBlock also tries to randomize which segment is
picked when there are overlapping segments to allow people to vote on
them. This is not relevant to us and we pick the segment with the
highest number of votes.

Improved information in exception traces.