Change file format
Changed the file format from csv -> json. This is because
csv does not directly support composite values (like the
tuple (start, end) that was being used before.

But the data that is recorded is also changed. Previously,
we used to store:
 - the full transcript
 - the part of the transcript that is sponsored content
 - the range of token (when split by whitespace) that is sponsored

Because we want to predct the start and end times of sponsor ranges
(that is what SponsorBlock needs after all), it makes sense to include
those in the dataset. Because the sponsored segment text was only a
subset of the full transcript, it didn't make sense to store a copy of
it - it can be extracted easily.

Because we might want to tweak the `error` threshold when picking
the start/end captions, it makes sense to do that when reading the
dataset - it is not an expensive step.

With those considered, the new format is:
 - `captions` - List of `{ start: sec, end: sec, text: str }`
 - `sponsor_range` - List of `{ start: sec, end: sec }`

The `sponsor_range` is the start and end time of each segment in the
video. The data is still chunked and so the file
format is `file.part.json.gz`.